requirements.txt
panda3d
panda3d-gltf
